fix: Progress indicator display + comprehensive Event Viewer integration plan
## Progress Indicator Fix
- Fix progress indicator showing empty values in template
- Convert percentage calculation to int to avoid float display issues
- Add template defaults for debugging (Step ? of ?, ?% complete)
- Ensure consistent progress calculation across welcome and basic_details views

## Event Viewer Integration Plan
- Comprehensive plan for registering all onboarding events in Event Viewer system
- Daily registration digest system with midday email alerts
- OSA compliance event logging (age verification, content restrictions)
- Security monitoring (suspicious registration patterns, fraud detection)
- Integration with existing ERROR_ALERTS_EMAIL_ADDRESS configuration

## Implementation Strategy
- Phase 1: Event logging integration (2-3 hours)
- Phase 2: Registration digest system (3-4 hours)
- Phase 3: Advanced monitoring & alerts (2-3 hours)
- Phase 4: Analytics dashboard enhancement (2 hours)

## Event Types Planned
- User registration started/completed events
- Onboarding step completion tracking
- Age verification compliance events
- Geographic user distribution
- Data validation failure monitoring
- Fraud detection and security alerts

## Email Alert System
- Daily digest at 12:00 PM with registration statistics
- Weekly trend comparison analytics
- Integration with existing email infrastructure
- GDPR-compliant event data collection
- Customizable alert thresholds for different event types

Ready for implementation of comprehensive registration monitoring.

+++ council_finance/views/onboarding_views.py
@@ -83,7 +83,7 @@ def welcome(request):
     # Calculate progress
     total_steps = 5 if is_uk_user else 4
     current_step = 1
-    progress_percent = round((current_step / total_steps) * 100)
+    progress_percent = int(round((current_step / total_steps) * 100))
     
     context = {
         'user_data': user_data,
@@ -146,7 +146,7 @@ def basic_details(request):
     # Calculate progress
     total_steps = 5 if profile.is_uk_user else 4
     current_step = 2
-    progress_percent = round((current_step / total_steps) * 100)
+    progress_percent = int(round((current_step / total_steps) * 100))
     
     context = {
         'form': form,
